Split repositories: one query per file with query prefix
Aligns the repo layer with the new "Database queries in own files"
project convention: each query function lives in its own
`query<Name>.server.ts` file under `app/repositories/`, and every
function is renamed with a `query` prefix.

- 5 monolithic repos → 25 single-purpose files.
- `find*` functions drop the verb and become `queryX`
  (e.g. findExerciseBySlug → queryExerciseBySlug).
- `create*`/`update*`/`delete*` keep the operation verb after the
  prefix (e.g. createExercise → queryCreateExercise).
- All 5 services updated to import the new files; data shapes and
  call sites are unchanged.

Verified: tsc, eslint, prettier all clean; 84/84 tests still pass.
