--- app/repositories/practiceSession.server.ts
@@ -1,238 +1,0 @@
-import type { Prisma } from '@prisma/client';
-import { db } from '~/utils/db.server';
-
-type SessionSectionItemInput = {
-  sectionId: string;
-  exerciseTypeId: string;
-  exerciseId?: string | null;
-  order: number;
-};
-
-type CreatePracticeSessionData = {
-  slug: string;
-  name?: string;
-  description?: string;
-  sessionLength: number;
-  seasonId?: string | null;
-  scheduledAt?: Date | null;
-  sectionItems: SessionSectionItemInput[];
-};
-
-export async function createPracticeSession(data: CreatePracticeSessionData) {
-  return db.practiceSession.create({
-    data: {
-      slug: data.slug,
-      name: data.name || null,
-      description: data.description || null,
-      sessionLength: data.sessionLength,
-      seasonId: data.seasonId ?? null,
-      scheduledAt: data.scheduledAt ?? null,
-      sectionItems: {
-        create: data.sectionItems,
-      },
-    },
-    include: {
-      sectionItems: {
-        include: {
-          section: {
-            include: {
-              translations: true,
-            },
-          },
-          exerciseType: {
-            include: {
-              translations: true,
-            },
-          },
-        },
-      },
-    },
-  });
-}
-
-export async function createPracticeSessionTx(
-  tx: Prisma.TransactionClient,
-  data: CreatePracticeSessionData
-) {
-  return tx.practiceSession.create({
-    data: {
-      slug: data.slug,
-      name: data.name || null,
-      description: data.description || null,
-      sessionLength: data.sessionLength,
-      seasonId: data.seasonId ?? null,
-      scheduledAt: data.scheduledAt ?? null,
-      sectionItems: {
-        create: data.sectionItems,
-      },
-    },
-  });
-}
-
-type FindPracticeSessionsOptions = {
-  standaloneOnly?: boolean;
-};
-
-export async function findAllPracticeSessions(options: FindPracticeSessionsOptions = {}) {
-  return db.practiceSession.findMany({
-    where: options.standaloneOnly ? { seasonId: null } : undefined,
-    orderBy: {
-      createdAt: 'desc',
-    },
-    include: {
-      _count: {
-        select: {
-          sectionItems: true,
-        },
-      },
-    },
-  });
-}
-
-export async function findPracticeSessionById(id: string, language: string) {
-  return db.practiceSession.findUnique({
-    where: { id },
-    include: {
-      season: {
-        select: { slug: true, name: true },
-      },
-      sectionItems: {
-        orderBy: [{ sectionId: 'asc' }, { order: 'asc' }],
-        include: {
-          section: {
-            include: {
-              translations: {
-                where: { language },
-              },
-            },
-          },
-          exerciseType: {
-            include: {
-              translations: {
-                where: { language },
-              },
-            },
-          },
-          exercise: {
-            select: { id: true, name: true, slug: true },
-          },
-        },
-      },
-    },
-  });
-}
-
-export async function findPracticeSessionBySlug(slug: string, language: string) {
-  return db.practiceSession.findUnique({
-    where: { slug },
-    include: {
-      season: {
-        select: { slug: true, name: true },
-      },
-      sectionItems: {
-        orderBy: [{ sectionId: 'asc' }, { order: 'asc' }],
-        include: {
-          section: {
-            include: {
-              translations: {
-                where: { language },
-              },
-            },
-          },
-          exerciseType: {
-            include: {
-              translations: {
-                where: { language },
-              },
-            },
-          },
-          exercise: {
-            select: { id: true, name: true, slug: true },
-          },
-        },
-      },
-    },
-  });
-}
-
-type UpdatePracticeSessionData = {
-  name?: string;
-  description?: string;
-  sessionLength: number;
-  scheduledAt?: Date | null;
-  sectionItems: Array<{
-    sectionId: string;
-    exerciseTypeId: string;
-    exerciseId?: string | null;
-    order: number;
-  }>;
-};
-
-export async function updatePracticeSession(id: string, data: UpdatePracticeSessionData) {
-  return db.$transaction(async tx => {
-    await tx.practiceSessionSectionItem.deleteMany({
-      where: { practiceSessionId: id },
-    });
-
-    return tx.practiceSession.update({
-      where: { id },
-      data: {
-        name: data.name || null,
-        description: data.description || null,
-        sessionLength: data.sessionLength,
-        scheduledAt: data.scheduledAt ?? null,
-        sectionItems: {
-          create: data.sectionItems,
-        },
-      },
-      include: {
-        sectionItems: {
-          include: {
-            section: {
-              include: {
-                translations: true,
-              },
-            },
-            exerciseType: {
-              include: {
-                translations: true,
-              },
-            },
-          },
-        },
-      },
-    });
-  });
-}
-
-export async function deletePracticeSession(id: string) {
-  return db.practiceSession.delete({
-    where: { id },
-  });
-}
-
-export async function findPracticeSessionsBySlugs(slugs: string[]) {
-  return db.practiceSession.findMany({
-    where: {
-      slug: {
-        in: slugs,
-      },
-    },
-    select: {
-      slug: true,
-    },
-  });
-}
-
-export async function findPracticeSessionsBySlugPrefix(prefix: string) {
-  return db.practiceSession.findMany({
-    where: {
-      slug: {
-        startsWith: prefix,
-      },
-    },
-    select: {
-      slug: true,
-    },
-  });
-}
