+++ app/repositories/queryDeletePracticeSession.server.ts
@@ -1,0 +1,7 @@
+import { db } from '~/utils/db.server';
+
+export async function queryDeletePracticeSession(id: string) {
+  return db.practiceSession.delete({
+    where: { id },
+  });
+}
