+++ app/repositories/queryDeleteSeason.server.ts
@@ -1,0 +1,7 @@
+import { db } from '~/utils/db.server';
+
+export async function queryDeleteSeason(id: string) {
+  return db.season.delete({
+    where: { id },
+  });
+}
