+++ app/repositories/queryExerciseById.server.ts
@@ -1,0 +1,7 @@
+import { db } from '~/utils/db.server';
+
+export function queryExerciseById(id: string) {
+  return db.exercise.findUnique({
+    where: { id },
+  });
+}
