+++ app/repositories/queryExerciseBySlug.server.ts
@@ -1,0 +1,7 @@
+import { db } from '~/utils/db.server';
+
+export function queryExerciseBySlug(slug: string) {
+  return db.exercise.findUnique({
+    where: { slug },
+  });
+}
