+++ app/repositories/queryExercisesBySlugs.server.ts
@@ -1,0 +1,14 @@
+import { db } from '~/utils/db.server';
+
+export async function queryExercisesBySlugs(slugs: string[]) {
+  return db.exercise.findMany({
+    where: {
+      slug: {
+        in: slugs,
+      },
+    },
+    select: {
+      slug: true,
+    },
+  });
+}
