+++ app/repositories/queryPracticeSessions.server.ts
@@ -1,0 +1,21 @@
+import { db } from '~/utils/db.server';
+
+export type QueryPracticeSessionsOptions = {
+  standaloneOnly?: boolean;
+};
+
+export async function queryPracticeSessions(options: QueryPracticeSessionsOptions = {}) {
+  return db.practiceSession.findMany({
+    where: options.standaloneOnly ? { seasonId: null } : undefined,
+    orderBy: {
+      createdAt: 'desc',
+    },
+    include: {
+      _count: {
+        select: {
+          sectionItems: true,
+        },
+      },
+    },
+  });
+}
