+++ app/repositories/queryRootExerciseTypesWithChildren.server.ts
@@ -1,34 +1,6 @@
 import { db } from '~/utils/db.server';
 
-export function findExerciseTypeWithHierarchy(exerciseTypeId: string, language: string) {
-  return db.exerciseType.findUnique({
-    where: { id: exerciseTypeId },
-    include: {
-      translations: {
-        where: { language },
-        select: { name: true },
-      },
-      parent: {
-        include: {
-          translations: {
-            where: { language },
-            select: { name: true },
-          },
-          parent: {
-            include: {
-              translations: {
-                where: { language },
-                select: { name: true },
-              },
-            },
-          },
-        },
-      },
-    },
-  });
-}
-
-export function findRootExerciseTypesWithChildren(language: string, groupSlug?: string) {
+export function queryRootExerciseTypesWithChildren(language: string, groupSlug?: string) {
   return db.exerciseType.findMany({
     where: {
       parentId: null,
